🔧 REFACTOR: Complete client portal blueprint migration
- Migrated all client portal authentication routes to client_portal blueprint
- Added client document upload and status update functionality
- Enhanced client portal with proper file handling and validation
- Fixed URL routing for client portal with proper redirects
- Cleaned up orphaned function implementations from app.py
- Progress: 81 routes migrated total (100 → 19 remaining)
